requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
transformers==4.33.2
faiss-cpu==1.7.4
pymongo==4.5.0
python-multipart==0.0.6
requests==2.31.0
Pillow==10.0.1